moved test functions to their appropriate place.
