+++ tests/unit/cli_test.py
@@ -4,56 +4,14 @@
 import logging
 import mock
 
-from datetime import datetime
-from hashlib import sha256
 
 from click.testing import CliRunner
+from datetime import datetime
 
 from storj import cli, model
 
+
 from .. import AbstractTestCase
-
-
-class FunctionsTestCase(AbstractTestCase):
-    """Test case for the package functions."""
-
-    @mock.patch('storj.cli.read_config')
-    def test_get_client(self, mock_read_config):
-        mock_read_config.return_value = {
-            cli.CFG_EMAIL: 'someone@example.com',
-            cli.CFG_PASSWORD: 'secret'
-        }
-
-        client = cli.get_client()
-
-        assert client is not None
-        assert client.email == 'someone@example.com'
-        assert client.password == sha256('secret'.encode('ascii')).hexdigest()
-
-        mock_read_config.assert_called_once_with()
-
-    @mock.patch('storj.cli.click.get_app_dir')
-    @mock.patch('storj.cli.configparser.RawConfigParser')
-    def test_read_config(self, mock_class_rawconfigparser, mock_app_dir):
-        # mock instance for ConfigParser.RawConfigParser
-        mrcp = mock_class_rawconfigparser.return_value
-        mrcp.read.return_value = None
-        mrcp.sections.return_value = ['storj']
-        mrcp.items.return_value = {
-            'email': 'someone@example.com',
-            'password': 'secret'
-        }.items()
-
-        mock_app_dir.return_value = '/nowhere'
-
-        assert cli.read_config() == {
-            cli.CFG_EMAIL: 'someone@example.com',
-            cli.CFG_PASSWORD: 'secret'
-        }
-
-        mrcp.read.assert_called_once_with(['/nowhere/storj.ini'])
-        mrcp.sections.assert_called_once_with()
-        mrcp.items.assert_called_once_with('storj')
 
 
 class BucketTestCase(AbstractTestCase):
