+++ tests/unit/storj_test.py
@@ -1,9 +1,11 @@
 # -*- coding: utf-8 -*-
 """Test cases for the storj package."""
 
+import mock
 import storj
 
 
+from hashlib import sha256
 from ecdsa import keys
 
 from .. import AbstractTestCase
@@ -25,3 +27,18 @@ class FunctionsTestCase(AbstractTestCase):
 
         self.assertEqual(key_pub, key.get_verifying_key())
         self.assertEqual(32, len(key.to_string()))
+
+    @mock.patch('storj.configuration.read_config')
+    def test_get_client(self, mock_read_config):
+        mock_read_config.return_value = {
+            storj.CFG_EMAIL: 'someone@example.com',
+            storj.CFG_PASSWORD: 'secret'
+        }
+
+        client = storj.get_client()
+
+        assert client is not None
+        assert client.email == 'someone@example.com'
+        assert client.password == sha256('secret'.encode('ascii')).hexdigest()
+
+        mock_read_config.assert_called_once_with()
